Finish implementing authn/authz
