+++ client/hooks/use_search_directions.ts
@@ -1,4 +1,3 @@
-import { redirect } from "next/dist/next-server/server/api-utils";
 import { useState } from "react";
 
 interface Direction {
@@ -14,24 +13,22 @@ export const useSearchDirections = (accessToken?: string) => {
   const searchDirections = async (origin: string, destinations: string[]) => {
     try {
       setIsLoading(true);
-      const response = await fetch(
-        `${
-          process.env.NEXT_PUBLIC_API_ENDPOINT
-        }?origin=${origin}&destinations=${destinations.join(",")}`,
-        {
-          headers: {
-            Authorization: `Bearer ${accessToken}`
-          }
-        }
-      );
+      const response = await fetch(process.env.NEXT_PUBLIC_API_ENDPOINT, {
+        method: "POST",
+        body: JSON.stringify({
+          function: "getDirections",
+          parameters: [origin, destinations.join(",")],
+        }),
+        headers: {
+          Authorization: `Bearer ${accessToken}`,
+        },
+      });
 
       const jsonResponse = await response.json();
-      setDirections(jsonResponse.body);
-    } 
-    catch (error)  {
+      setDirections(jsonResponse.response.result.body);
+    } catch (error) {
       console.error(error);
-    }
-    finally{
+    } finally {
       setIsLoading(false);
     }
   };
