+++ client/pages/api/auth/[...nextauth].ts
@@ -25,7 +25,7 @@ const options: InitOptions = {
     async session(session, user) {
 	  session.accessToken = (user as any).accessToken;
       return session;
-    },
+	},
   },
 };
